requirements-chemistry.txt
bespokelabs-curator>=1.0.0
datasets>=2.14.0
pydantic>=2.0.0
rdkit-pypi>=2022.9.5
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
